requirements.txt
pinecone-client
fiftyone
torch